move 'reduceRight' tests to separate file
