+++ types/ramda/ramda-tests.ts
@@ -134,16 +134,6 @@ function i(x: number) {
 R.times(i, 5);
 
 (() => {
-    const pairs = [["a", 1], ["b", 2], ["c", 3]];
-
-    function flattenPairs(pair: [string, number], acc: Array<string|number>): Array<string|number> {
-        return acc.concat(pair);
-    }
-
-    R.reduceRight(flattenPairs, [], pairs); // => [ 'c', 3, 'b', 2, 'a', 1 ]
-})();
-
-(() => {
     function isOdd(n: number) {
         return n % 2 === 1;
     }
@@ -185,24 +175,6 @@ R.times(i, 5);
     const headLens = R.lensIndex(0);
     R.view(headLens, ["a", "b", "c"]);            // => 'a'
     R.set(headLens, "x", ["a", "b", "c"]);        // => ['x', 'b', 'c']
-};
-
-interface KeyValuePair<K, V> extends Array<K | V> {
-    0: K;
-    1: V;
-}
-type Pair = KeyValuePair<string, number>;
-
-() => {
-    const pairs: Pair[] = [["a", 1], ["b", 2], ["c", 3]];
-
-    function flattenPairs(pair: Pair, acc: Array<string|number>): Array<string|number> {
-        return acc.concat(pair);
-    }
-
-    R.reduceRight(flattenPairs, [], pairs); // => [ 'c', 3, 'b', 2, 'a', 1 ]
-    R.reduceRight(flattenPairs, [])(pairs); // => [ 'c', 3, 'b', 2, 'a', 1 ]
-    R.reduceRight(flattenPairs)([], pairs); // => [ 'c', 3, 'b', 2, 'a', 1 ]
 };
 
 () => {
