+++ types/ramda/tsconfig.json
@@ -188,6 +188,7 @@
         "test/propSatisfies-tests.ts",
         "test/range-tests.ts",
         "test/reduce-tests.ts",
-        "test/reduceBy-tests.ts"
+        "test/reduceBy-tests.ts",
+        "test/reduceRight-tests.ts"
     ]
 }
